requirements.txt
Flask
pandas
joblib
tensorflow
scikit-learn
plotly
paho-mqtt
requests
twilio
keras
